requirements.txt
bleach==1.5.0
boto3
botocore
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.28.5
decorator==4.3.0
docutils==0.14
enum34==1.1.6
Flask==1.0.2
future==0.16.0
glfw==1.7.0
gym==0.7.4
html5lib==0.9999999
idna==2.7
imageio==2.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2>=2.10.1
jmespath==0.9.3
joblib==0.12.4
jsonschema==2.6.0
kiwisolver==1.0.1
matplotlib==2.2.3
moviepy==0.2.3.5
mpi4py==3.0.0
mujoco-py==0.5.7
nbformat==4.4.0
nose==1.3.7
numpy==1.15.1
Pillow==5.2.0
pip==10.0.1
plotly==3.2.0
protobuf==3.6.1
pycparser==2.18
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.2.0
PyPrind==2.11.2
pyquaternion==0.9.2
python-dateutil==2.7.3
pytz==2018.5
requests>=2.20.0
retrying==1.3.3
s3transfer==0.1.13
scipy==1.1.0
setuptools==40.2.0
six==1.11.0
tensorflow-gpu==1.13.1
tensorflow-tensorboard==0.4.0
tqdm==4.25.0
traitlets==4.3.2
urllib3>=1.24.2
Werkzeug==0.14.1
wheel==0.31.1
